move operators and interpolation into cellCentred, cleanup includes
