+++ include/NeoFOAM/finiteVolume/cellCentred/interpolation/upwind.hpp
@@ -5,7 +5,7 @@
 
 #include "NeoFOAM/fields/field.hpp"
 #include "NeoFOAM/core/executor/executor.hpp"
-#include "NeoFOAM/finiteVolume/interpolation/surfaceInterpolation.hpp"
+#include "NeoFOAM/finiteVolume/cellCentred/interpolation/surfaceInterpolation.hpp"
 #include "NeoFOAM/mesh/unstructured.hpp"
 #include "NeoFOAM/mesh/stencil/fvccGeometryScheme.hpp"
