+++ include/NeoFOAM/finiteVolume/cellCentred/operators/gaussGreenDiv.hpp
@@ -3,18 +3,14 @@
 
 #pragma once
 
-#include "NeoFOAM/fields/field.hpp"
-#include "NeoFOAM/core/executor/executor.hpp"
-
-#include "NeoFOAM/mesh/unstructured.hpp"
-#include "NeoFOAM/finiteVolume/cellCentred.hpp"
-
 #include "Kokkos_Core.hpp"
+
 #include <functional>
 
-#include "NeoFOAM/finiteVolume/interpolation/linear.hpp"
-#include "NeoFOAM/finiteVolume/interpolation/upwind.hpp"
-#include "NeoFOAM/finiteVolume/interpolation/surfaceInterpolation.hpp"
+#include "NeoFOAM/fields/field.hpp"
+#include "NeoFOAM/core/executor/executor.hpp"
+#include "NeoFOAM/mesh/unstructured.hpp"
+#include "NeoFOAM/finiteVolume/cellCentred/interpolation/surfaceInterpolation.hpp"
 
 namespace NeoFOAM
 {
