+++ include/NeoFOAM/finiteVolume/cellCentred/operators/gaussGreenGrad.hpp
@@ -1,19 +1,16 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2023 NeoFOAM authors
+
 #pragma once
+
+#include <functional>
+
+#include "Kokkos_Core.hpp"
 
 #include "NeoFOAM/fields/field.hpp"
 #include "NeoFOAM/core/executor/executor.hpp"
 #include "NeoFOAM/mesh/unstructured.hpp"
-
-#include "NeoFOAM/finiteVolume/cellCentred.hpp"
-
-#include "Kokkos_Core.hpp"
-#include <functional>
-#include "NeoFOAM/finiteVolume/interpolation/linear.hpp"
-#include "NeoFOAM/finiteVolume/interpolation/upwind.hpp"
-#include "NeoFOAM/finiteVolume/interpolation/surfaceInterpolation.hpp"
-
+#include "NeoFOAM/finiteVolume/cellCentred/interpolation/surfaceInterpolation.hpp"
 
 namespace NeoFOAM
 {
