+++ src/finiteVolume/cellCentred/interpolation/linear.cpp
@@ -1,8 +1,9 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2023 NeoFOAM authors
 
-#include "NeoFOAM/finiteVolume/interpolation/linear.hpp"
 #include <memory>
+
+#include "NeoFOAM/finiteVolume/cellCentred/interpolation/linear.hpp"
 #include "NeoFOAM/core/error.hpp"
 #include "NeoFOAM/core/parallelAlgorithms.hpp"
