+++ src/finiteVolume/cellCentred/operators/gaussGreenDiv.cpp
@@ -1,9 +1,12 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2023 NeoFOAM authors
 
-#include "NeoFOAM/finiteVolume/operators/gaussGreenDiv.hpp"
 #include <functional>
+
+#include "NeoFOAM/finiteVolume/cellCentred/operators/gaussGreenDiv.hpp"
 #include "NeoFOAM/core/parallelAlgorithms.hpp"
+#include "NeoFOAM/mesh/stencil/fvccGeometryScheme.hpp"
+
 namespace NeoFOAM
 {
